requirements.txt
colorlog==6.7.0
homeassistant==2023.5.4
pip>=21.0,<23.2
ruff==0.0.267
pre-commit>=2.11.1
discord.py